removes all old submodules
