+++ imagepypelines/core/CameraCapture.py
