+++ imagepypelines/core/ConfigFactory.py
