+++ imagepypelines/core/Emailer.py
