+++ imagepypelines/core/ImageWriter.py
