+++ imagepypelines/core/Printer.py
@@ -107,6 +107,40 @@ def set_global_printout_level(log_level):
 
     Printer.GLOBAL_LOG_LEVEL = log_level
 
+def get_default_printer():
+    """gets the imagepypelines default printer"""
+    return get_printer('imagepypelines')
+
+def debug(*messages):
+    """prints a 'debug' level message to the imagepypelines default printer"""
+    printer = get_default_printer()
+    printer.debug(*messages)
+
+def info(*messages):
+    """prints a 'info' level message to the imagepypelines default printer"""
+    printer = get_default_printer()
+    printer.info(*messages)
+
+def warning(*messages):
+    """prints a 'warning' level message to the imagepypelines default printer"""
+    printer = get_default_printer()
+    printer.warning(*messages)
+
+def error(*messages):
+    """prints a 'error' level message to the imagepypelines default printer"""
+    printer = get_default_printer()
+    printer.error(*messages)
+
+def critical(*messages):
+    """prints a 'critical' level message to the imagepypelines default printer"""
+    printer = get_default_printer()
+    printer.critical(*messages)
+
+def comment(*messages):
+    """prints a 'comment' level message to the imagepypelines default printer"""
+    printer = get_default_printer()
+    printer.comment(*messages)
+    
 
 class Printer(object):
     """
