+++ imagepypelines/core/VideoWriter.py
