+++ imagepypelines/core/format.py
