+++ imagepypelines/core/output.py
