+++ imagepypelines/core/visualbagofwords.py
